requirements.txt
scikit-learn~=1.4.1.post1
matplotlib~=3.8.3
numpy~=1.26.4
umap~=0.1.1
pandas~=2.2.0
scipy~=1.12.0
tqdm~=4.66.2
setuptools~=59.6.0
requests~=2.31.0
beautifulsoup4~=4.12.3
pillow~=10.2.0
tensorflow~=2.15.0.post1
dash~=2.16.1
plotly~=5.20.0
keras~=2.15.0